requirements.txt
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg 
mysqlclient